+++ lib/rpcc_dpdk/tapi_ethdev.c
@@ -38,7 +38,7 @@
 
 
 struct tarpc_rte_eth_conf *
-tapi_rpc_rte_eth_make_eth_conf(rcf_rpc_server *rpcs, uint8_t port_id,
+tapi_rpc_rte_eth_make_eth_conf(rcf_rpc_server *rpcs, uint16_t port_id,
                                struct tarpc_rte_eth_conf *eth_conf)
 {
     UNUSED(rpcs);
@@ -52,7 +52,7 @@ tapi_rpc_rte_eth_make_eth_conf(rcf_rpc_server *rpcs, uint8_t port_id,
 }
 
 int
-tapi_rpc_rte_eth_dev_configure_def(rcf_rpc_server *rpcs, uint8_t port_id,
+tapi_rpc_rte_eth_dev_configure_def(rcf_rpc_server *rpcs, uint16_t port_id,
                                    uint16_t nb_rx_queue, uint16_t nb_tx_queue)
 {
     struct tarpc_rte_eth_conf   eth_conf;
@@ -63,7 +63,7 @@ tapi_rpc_rte_eth_dev_configure_def(rcf_rpc_server *rpcs, uint8_t port_id,
 
 /* See description in tapi_rpc_rte_ethdev.h */
 te_errno
-tapi_rpc_add_mac_as_octstring2kvpair(rcf_rpc_server *rpcs, uint8_t port_id,
+tapi_rpc_add_mac_as_octstring2kvpair(rcf_rpc_server *rpcs, uint16_t port_id,
                                      te_kvpair_h *head, const char *name)
 {
     struct tarpc_ether_addr     mac_addr;
